requirements.txt
python-dotenv
requests
google-cloud-aiplatform
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
playwright
beautifulsoup4
lxml
pandas
urllib3
dnspython